split theme in common theme things and plalette files
