+++ src/themes/types.ts
@@ -1,0 +1,98 @@
+import { PaletteColor, TypeText, PaletteColorOptions } from '@mui/material';
+
+declare module '@mui/material/styles' {
+    interface TypeBackground {
+        overPaper?: string;
+    }
+
+    interface BlurStrength {
+        base: string;
+        muted: string;
+        faint: string;
+    }
+    interface BlurStrengthOptions {
+        base?: string;
+        muted?: string;
+        faint?: string;
+    }
+
+    interface FixedColor {
+        white: string;
+        black: string;
+        strokeMutedWhite: string;
+    }
+    interface Palette {
+        accent: PaletteColor;
+        fill: PaletteColor;
+        backdrop: PaletteColor;
+        blur: BlurStrength;
+        danger: PaletteColor;
+        stroke: TypeText;
+        fixed: FixedColor;
+    }
+    interface PaletteOptions {
+        accent?: PaletteColorOptions;
+        danger?: PaletteColorOptions;
+        fill?: PaletteColorOptions;
+        backdrop?: PaletteColorOptions;
+        blur?: BlurStrengthOptions;
+        stroke?: Partial<TypeText>;
+        fixed?: Partial<FixedColor>;
+    }
+
+    interface TypographyVariants {
+        title: React.CSSProperties;
+        subtitle: React.CSSProperties;
+        mini: React.CSSProperties;
+    }
+
+    interface TypographyVariantsOptions {
+        title?: React.CSSProperties;
+        subtitle?: React.CSSProperties;
+        mini?: React.CSSProperties;
+    }
+}
+
+declare module '@mui/material/Button' {
+    export interface ButtonPropsColorOverrides {
+        accent: true;
+        danger: true;
+    }
+}
+declare module '@mui/material/Checkbox' {
+    export interface CheckboxPropsColorOverrides {
+        accent: true;
+    }
+}
+
+declare module '@mui/material/Typography' {
+    interface TypographyPropsVariantOverrides {
+        title: true;
+        subtitle: true;
+    }
+}
+
+declare module '@mui/material/Switch' {
+    interface SwitchPropsColorOverrides {
+        accent: true;
+    }
+}
+
+declare module '@mui/material/SvgIcon' {
+    interface SvgIconPropsColorOverrides {
+        accent: true;
+    }
+}
+
+declare module '@mui/material/Alert' {
+    export interface AlertPropsColorOverrides {
+        accent: true;
+    }
+}
+
+declare module '@mui/material/CircularProgress' {
+    export interface CircularProgressPropsColorOverrides {
+        accent: true;
+    }
+}
+export {};
